Make the name of state files dynamic

+++ src/Listener.py
@@ -1,80 +1,80 @@
 # pylint: disable=bad-whitespace,bad-continuation,line-too-long,multiple-statements,trailing-whitespace,trailing-newlines
 """PCM
-    PCM Server
+    PCM Listener
 """
 
 import sys
 import asyncio
-from os import getpid, kill, unlink, _exit
+from os import getpid, kill, unlink, _exit, path
 from signal import signal, SIGTERM, SIGUSR1
 from time import sleep, time
 from pickle import dumps,loads
+from psutil import process_iter
 
 import PCM
 
-class Server:
+class Listener:
 
     def __init__(self, configuration, logger):
     
-        self.pidfile = configuration['server_pid_file']
         self.logger = logger
         self.state = {}
         self.state['startTime'] = None
         self.state['uptime'] = 0
-        self.state['filename'] = configuration['server_state_file']
+        self.state['filename'] = configuration['pcm_dir']+'/'+'listener.state'
+
+    def process_exists(self):
+                
+        cmdline = [path.basename(sys.executable),sys.argv[0],'listener']
+        for p in process_iter():
+            if p.cmdline() == cmdline: return p.pid
+        return False
 
     def start(self):
 
-        if self.status(): return False
-        print('Server starts. PID={}'.format(str(getpid())), file=sys.stderr)
+        print('Listener starts. PID={}'.format(str(getpid())), file=sys.stderr)
+        
         try:
             self.loop = asyncio.get_event_loop()
             self.coro = asyncio.start_server(self.handle_request, '127.0.0.1', 1331, loop=self.loop)
             self.server = self.loop.run_until_complete(self.coro)
             self.logger.log('Serving on {}'.format(self.server.sockets[0].getsockname()),0)
             self.state['startTime'] = int(time())
-            with open(self.pidfile,'w') as f: f.write(str(getpid()))
             with open(self.state['filename'] , 'wb') as f:
                 f.write(dumps(self.state))
             signal(SIGUSR1, self.close_server)
-            message = 'Server is listening on {}. PID={}'.format(self.server.sockets[0].getsockname(),getpid())
+            message = 'Listening on {}. PID={}'.format(self.server.sockets[0].getsockname(),getpid())
             print(message, file=sys.stderr)
             self.logger.log(message, 1)
             self.loop.run_forever()
+            
         except OSError as err:
             print(str(err), file=sys.stderr)
 
     def stop(self):
-        try:
-            with open(self.pidfile,'r') as f: pid = int(f.readline())
-            kill(pid,SIGUSR1)
-            unlink(self.pidfile)
+        
+        pid = self.process_exists()
+        
+        if pid:
+            kill(pid,SIGTERM)
             unlink(self.state['filename'])
-            print('Server stopped.', file=sys.stderr)
-        except (AttributeError,FileNotFoundError,ProcessLookupError): 
-            print('Server not running.', file=sys.stderr)
+            print('Listener stopped.', file=sys.stderr)
+            self.logger.log('Listener PID='+str(pid)+' stopped.', 1)
+            return True
+        else:
+            print('Listener is not running.', file=sys.stderr)
+            return False
  
     def status(self):
-        try:
-            with open(self.pidfile,'r') as f: pid = int(f.readline())
-            try:
-                with open(self.state['filename'], 'rb') as f:
-                    self.state = loads(f.read())
-                    self.state['uptime'] = int(time()) - self.state['startTime']
-            except (FileNotFoundError, EOFError) as e: pass
-            with open(self.state['filename'], 'wb') as f:
-                f.write(dumps(self.state))
-            print('Server is running. PID={} Uptime={}'.format(str(pid), str(self.state['uptime'])), file=sys.stderr)
+        
+        pid = self.process_exists()
+        
+        if pid:
+            print('Listener is running. PID={}'.format(str(pid)), file=sys.stderr)
             return True
-        except (AttributeError,OSError) as err:
-            try:
-                if err.errno == errno.EPERM:
-                    print('Server is running but access is denied. PID='+str(pid), file=sys.stderr)
-                    return False
-                else: return False
-            except NameError as ne:
-                print('Server not running.', file=sys.stderr)
-                return False
+        else:
+            print('Listener is not running.', file=sys.stderr)
+            return False
         
     def close_server(self, signum, frame):
         self.logger.log('Server PID='+str(getpid())+' received signal '+str(signum)+' ('+str(frame)+'). Stopping…', 1)
